+++ src/components/BodyTeleop/index.jsx
@@ -14,9 +14,10 @@ import MicOff from '@material-ui/icons/MicOff';
 
 import Colors from '../../colors';
 import { deviceNamePretty } from '../../utils';
-import { isMobile, isChrome, isFirefox, isSafari, isIos } from '../../utils/browser';
-import { BodyTeleopConnection, checkSslTrust, getDeviceBaseUrl } from '../../utils/bodyteleop';
+import { BodyTeleopConnection } from '../../utils/bodyteleop';
 import { ArrowBackBold } from '../../icons';
+
+import styles from "./styles"
 
 const QUICK_SOUNDS = [
   { key: 'engage', label: 'Engage' },
@@ -29,829 +30,6 @@ const CAMERAS = [
   { key: 'wideRoad', label: 'road', num: '1' },
   { key: 'driver', label: 'driver', num: '2' },
 ];
-
-const styles = () => ({
-  root: {
-    position: 'fixed',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    zIndex: 1300,
-    background: Colors.black,
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  header: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: '8px 12px',
-    background: Colors.grey950,
-    borderBottom: `1px solid ${Colors.white10}`,
-    minHeight: 48,
-    zIndex: 10,
-  },
-  headerTitle: {
-    fontSize: 16,
-    fontWeight: 500,
-    marginLeft: 8,
-    flex: 1,
-  },
-  backButton: {
-    color: Colors.white,
-    padding: 8,
-  },
-  videoContainer: {
-    flex: 1,
-    position: 'relative',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    overflow: 'hidden',
-    background: Colors.black,
-  },
-  video: {
-    width: '100%',
-    height: '100%',
-    objectFit: 'contain',
-  },
-  hudTopRight: {
-    position: 'absolute',
-    top: 12,
-    right: 12,
-    zIndex: 10,
-    display: 'flex',
-    alignItems: 'center',
-    gap: '8px',
-  },
-  hudBottomCenter: {
-    position: 'absolute',
-    bottom: 8,
-    left: '50%',
-    transform: 'translateX(-50%)',
-    zIndex: 10,
-  },
-  hudPill: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: '6px',
-    height: 28,
-    padding: '0 10px',
-    borderRadius: 14,
-    background: 'rgba(0,0,0,0.5)',
-    backdropFilter: 'blur(8px)',
-    border: `1px solid ${Colors.white10}`,
-  },
-  hudText: {
-    fontSize: 12,
-    color: Colors.white70,
-  },
-  statusDot: {
-    width: 8,
-    height: 8,
-    borderRadius: '50%',
-  },
-  connectOverlay: {
-    position: 'absolute',
-    inset: 0,
-    zIndex: 10,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  connectContent: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    gap: '12px',
-  },
-  connectButton: {
-    padding: '10px 24px',
-    borderRadius: 24,
-    color: Colors.white,
-    fontSize: 14,
-    fontWeight: 500,
-    background: 'rgba(255,255,255,0.15)',
-    boxShadow: '0 4px 24px rgba(0,0,0,0.4)',
-    border: `1px solid ${Colors.white20}`,
-    backdropFilter: 'blur(12px)',
-    cursor: 'pointer',
-    textTransform: 'none',
-    '& span': {
-      display: 'flex',
-      alignItems: 'center',
-      gap: '8px',
-    },
-    '&:hover': {
-      background: 'rgba(255,255,255,0.25)',
-    },
-  },
-  connectButtonFailed: {
-    background: 'rgba(220,38,38,0.6)',
-    '&:hover': {
-      background: 'rgba(220,38,38,0.7)',
-    },
-  },
-  connectButtonConnecting: {
-    animation: '$pulse 1.5s ease-in-out infinite',
-  },
-  progressBar: {
-    width: 240,
-    height: 4,
-    borderRadius: 2,
-    background: 'rgba(255,255,255,0.1)',
-    overflow: 'hidden',
-  },
-  progressFill: {
-    height: '100%',
-    borderRadius: 2,
-    background: 'rgba(255,255,255,0.7)',
-    transition: 'width 0.4s ease',
-  },
-  progressLabel: {
-    fontSize: 12,
-    color: 'rgba(255,255,255,0.5)',
-  },
-  errorBanner: {
-    maxWidth: 280,
-    borderRadius: 8,
-    padding: '6px 12px',
-    textAlign: 'center',
-    fontSize: 12,
-    color: '#fca5a5',
-    background: 'rgba(220,38,38,0.4)',
-    backdropFilter: 'blur(8px)',
-  },
-  infoBanner: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '10px',
-    borderRadius: 20,
-    padding: '6px 12px',
-    fontSize: 12,
-    color: Colors.white50,
-    background: 'rgba(0,0,0,0.3)',
-    backdropFilter: 'blur(8px)',
-  },
-  // Joystick
-  joystickArea: {
-    position: 'absolute',
-    bottom: 16,
-    right: 16,
-    zIndex: 10,
-    width: 160,
-    height: 160,
-    borderRadius: '50%',
-    background: 'radial-gradient(circle at 30% 30%, rgba(255,255,255,0.25), rgba(255,255,255,0.05))',
-    boxShadow: 'inset 0 0 20px rgba(255,255,255,0.1), 0 4px 20px rgba(0,0,0,0.4)',
-    border: '1.5px solid rgba(255,255,255,0.2)',
-    backdropFilter: 'blur(12px)',
-    touchAction: 'none',
-    '@media (min-width: 768px)': {
-      width: 160,
-      height: 160,
-    },
-  },
-  joystickAreaMobile: {
-    position: 'relative',
-    bottom: 'auto',
-    right: 'auto',
-    width: 'auto',
-    height: '100%',
-    aspectRatio: '1 / 1',
-    maxWidth: '100%',
-    background: 'radial-gradient(circle at 30% 30%, rgba(255,255,255,0.25), rgba(255,255,255,0.05))',
-    boxShadow: 'inset 0 0 20px rgba(255,255,255,0.1), 0 4px 20px rgba(0,0,0,0.4)',
-    border: '1.5px solid rgba(255,255,255,0.2)',
-    backdropFilter: 'blur(12px)',
-    touchAction: 'none',
-  },
-  joystickAreaSquare: {
-    borderRadius: 16,
-    background: 'linear-gradient(135deg, rgba(255,255,255,0.2), rgba(255,255,255,0.05))',
-  },
-  joystickCrosshairV: {
-    position: 'absolute',
-    left: '50%',
-    top: 8,
-    bottom: 8,
-    width: 1,
-    transform: 'translateX(-50%)',
-    background: 'rgba(255,255,255,0.1)',
-  },
-  joystickCrosshairH: {
-    position: 'absolute',
-    top: '50%',
-    left: 8,
-    right: 8,
-    height: 1,
-    transform: 'translateY(-50%)',
-    background: 'rgba(255,255,255,0.1)',
-  },
-  joystickCenter: {
-    position: 'absolute',
-    left: '50%',
-    top: '50%',
-    width: 6,
-    height: 6,
-    transform: 'translate(-50%, -50%)',
-    borderRadius: '50%',
-    background: 'rgba(255,255,255,0.3)',
-  },
-  joystickThumb: {
-    position: 'absolute',
-    width: 52,
-    height: 52,
-    borderRadius: '50%',
-    transform: 'translate(-50%, -50%)',
-    background: 'radial-gradient(circle at 35% 35%, rgba(255,255,255,0.4), rgba(255,255,255,0.1))',
-    boxShadow: 'inset 0 1px 4px rgba(255,255,255,0.3), 0 2px 8px rgba(0,0,0,0.3)',
-    border: '1px solid rgba(255,255,255,0.25)',
-    willChange: 'left, top',
-    '@media (min-width: 768px)': {
-      width: 56,
-      height: 56,
-    },
-  },
-  joystickThumbActive: {
-    background: 'radial-gradient(circle at 35% 35%, rgba(255,255,255,0.6), rgba(255,255,255,0.15))',
-    boxShadow: 'inset 0 1px 4px rgba(255,255,255,0.4), 0 2px 12px rgba(0,0,0,0.4)',
-    border: '1px solid rgba(255,255,255,0.35)',
-  },
-  // WASD keys
-  wasdContainer: {
-    position: 'absolute',
-    bottom: 16,
-    left: 16,
-    zIndex: 10,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    gap: '2px',
-    borderRadius: 12,
-    padding: 8,
-    background: 'rgba(0,0,0,0.3)',
-    backdropFilter: 'blur(8px)',
-  },
-  wasdRow: {
-    display: 'flex',
-    gap: '2px',
-  },
-  wasdKey: {
-    width: 36,
-    height: 36,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 6,
-    fontSize: 12,
-    fontWeight: 700,
-    cursor: 'pointer',
-    userSelect: 'none',
-    transition: 'background 0.15s, color 0.15s',
-  },
-  wasdKeyInactive: {
-    background: 'rgba(255,255,255,0.1)',
-    color: 'rgba(255,255,255,0.6)',
-  },
-  wasdKeyActive: {
-    background: 'rgba(255,255,255,0.3)',
-    color: Colors.white,
-  },
-  // Controls group (camera + sounds)
-  controlsGroup: {
-    position: 'absolute',
-    bottom: 16,
-    left: 16,
-    zIndex: 10,
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'stretch',
-    gap: '10px',
-    borderRadius: 14,
-    padding: 8,
-    background: 'rgba(0,0,0,0.35)',
-    backdropFilter: 'blur(8px)',
-  },
-  controlsGroupPortrait: {
-    position: 'relative',
-    bottom: 'auto',
-    left: 'auto',
-    transform: 'none',
-    alignSelf: 'stretch',
-    borderRadius: 0,
-    flexShrink: 0,
-    flexDirection: 'column',
-    alignItems: 'stretch',
-    gap: '6px',
-  },
-  controlsColumn: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    gap: '5px',
-  },
-  controlsRow: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '4px',
-  },
-  portraitRow: {
-    display: 'flex',
-    alignItems: 'flex-end',
-    gap: '6px',
-  },
-  portraitCategory: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '3px',
-  },
-  controlsLabel: {
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    textTransform: 'uppercase',
-    color: 'rgba(255,255,255,0.35)',
-    minWidth: 52,
-    flexShrink: 0,
-  },
-  controlsLabelPortrait: {
-    fontSize: 8,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    textTransform: 'uppercase',
-    color: 'rgba(255,255,255,0.35)',
-    minWidth: 36,
-    flexShrink: 0,
-  },
-  controlsLabelBelow: {
-    fontSize: 9,
-    fontWeight: 600,
-    letterSpacing: 0.5,
-    textTransform: 'uppercase',
-    color: 'rgba(255,255,255,0.35)',
-    textAlign: 'center',
-    lineHeight: 1,
-  },
-  controlsButtons: {
-    display: 'flex',
-    gap: '3px',
-    alignItems: 'center',
-  },
-  cameraButton: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '6px',
-    height: 32,
-    padding: '0 10px',
-    borderRadius: 8,
-    fontSize: 11,
-    fontWeight: 600,
-    cursor: 'pointer',
-    userSelect: 'none',
-    transition: 'background 0.15s, color 0.15s',
-  },
-  cameraButtonInactive: {
-    background: 'rgba(255,255,255,0.1)',
-    color: 'rgba(255,255,255,0.6)',
-  },
-  cameraButtonActive: {
-    background: 'rgba(255,255,255,0.3)',
-    color: Colors.white,
-  },
-  cameraButtonPortrait: {
-    width: 28,
-    height: 28,
-    padding: 0,
-    fontSize: 11,
-    justifyContent: 'center',
-  },
-  cameraButtonKey: {
-    width: 18,
-    height: 18,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 4,
-    fontSize: 10,
-    fontWeight: 700,
-    background: 'rgba(255,255,255,0.15)',
-  },
-  soundButton: {
-    height: 32,
-    padding: '0 10px',
-    borderRadius: 8,
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 0.2,
-    textTransform: 'uppercase',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    color: 'rgba(255,255,255,0.7)',
-    background: 'rgba(255,255,255,0.12)',
-    border: '1px solid rgba(255,255,255,0.12)',
-    cursor: 'pointer',
-    userSelect: 'none',
-    '&:hover': {
-      color: Colors.white,
-      background: 'rgba(255,255,255,0.2)',
-    },
-  },
-  soundButtonPortrait: {
-    height: 28,
-    padding: '0 8px',
-    fontSize: 9,
-  },
-  // Portrait mode
-  portraitContent: {
-    padding: 16,
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '12px',
-  },
-  statusRow: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    borderRadius: 8,
-    background: Colors.grey900,
-    padding: 12,
-  },
-  statusLeft: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '10px',
-  },
-  batteryPill: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '6px',
-    fontSize: 14,
-  },
-  portraitButton: {
-    borderRadius: 24,
-    padding: '10px 20px',
-    textTransform: 'none',
-    fontWeight: 500,
-  },
-  infoBox: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: '8px',
-    borderRadius: 8,
-    background: Colors.grey900,
-    padding: 12,
-    fontSize: 12,
-    color: Colors.white50,
-  },
-  infoRow: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: '10px',
-  },
-  statsToggle: {
-    position: 'absolute',
-    top: 12,
-    left: '50%',
-    transform: 'translateX(-50%)',
-    zIndex: 10,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  statsTogglePortrait: {
-    position: 'absolute',
-    top: 0,
-    right: 4,
-    zIndex: 10,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'flex-end',
-  },
-  statsToggleButton: {
-    height: 28,
-    padding: '0 10px',
-    borderRadius: 14,
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 1,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    color: 'rgba(255,255,255,0.6)',
-    background: 'rgba(0,0,0,0.4)',
-    backdropFilter: 'blur(8px)',
-    border: '1px solid rgba(255,255,255,0.1)',
-    cursor: 'pointer',
-    userSelect: 'none',
-    '&:hover': {
-      color: 'rgba(255,255,255,0.9)',
-      background: 'rgba(0,0,0,0.6)',
-    },
-  },
-  statsPanel: {
-    marginTop: 2,
-    padding: '3px 6px',
-    borderRadius: 5,
-    background: 'rgba(0,0,0,0.7)',
-    backdropFilter: 'blur(12px)',
-    border: '1px solid rgba(255,255,255,0.1)',
-    minWidth: 120,
-    fontFamily: 'monospace',
-    '@media (min-width: 768px)': {
-      padding: '10px 16px',
-      minWidth: 240,
-      borderRadius: 10,
-    },
-  },
-  statsRow: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    lineHeight: 1.2,
-    '@media (min-width: 768px)': {
-      padding: '3px 0',
-    },
-  },
-  statsLabel: {
-    fontSize: 8,
-    color: 'rgba(255,255,255,0.45)',
-    marginRight: 6,
-    '@media (min-width: 768px)': {
-      fontSize: 13,
-      marginRight: 18,
-    },
-  },
-  statsValue: {
-    fontSize: 8,
-    color: 'rgba(255,255,255,0.85)',
-    textAlign: 'right',
-    '@media (min-width: 768px)': {
-      fontSize: 13,
-    },
-  },
-  statsDivider: {
-    height: 1,
-    background: 'rgba(255,255,255,0.08)',
-    margin: '1px 0',
-    '@media (min-width: 768px)': {
-      margin: '5px 0',
-    },
-  },
-  audioLevelRow: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: '0px 0',
-    gap: '3px',
-    '@media (min-width: 768px)': {
-      padding: '2px 0',
-      gap: '6px',
-    },
-  },
-  audioLevelIcon: {
-    fontSize: 8,
-    color: 'rgba(255,255,255,0.45)',
-    flexShrink: 0,
-    width: 10,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  audioLevelTrack: {
-    flex: 1,
-    height: 3,
-    borderRadius: 2,
-    background: 'rgba(255,255,255,0.08)',
-    overflow: 'hidden',
-  },
-  audioLevelFill: {
-    height: '100%',
-    borderRadius: 2,
-    transition: 'width 80ms linear',
-  },
-  latencyGraph: {
-    width: '100%',
-    height: 30,
-    marginTop: 1,
-    borderRadius: 3,
-    background: 'rgba(0,0,0,0.3)',
-    '@media (min-width: 768px)': {
-      height: 90,
-      marginTop: 4,
-      borderRadius: 6,
-    },
-  },
-  latencySectionHeader: {
-    fontSize: 7,
-    fontWeight: 700,
-    color: 'rgba(255,255,255,0.35)',
-    letterSpacing: 0.5,
-    lineHeight: 1.2,
-    '@media (min-width: 768px)': {
-      fontSize: 11,
-    },
-    padding: '2px 0 1px',
-  },
-  screenshotButton: {
-    width: 28,
-    height: 28,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 14,
-    background: 'rgba(0,0,0,0.5)',
-    backdropFilter: 'blur(8px)',
-    border: `1px solid ${Colors.white10}`,
-    color: Colors.white70,
-    cursor: 'pointer',
-    '&:hover': {
-      background: 'rgba(0,0,0,0.7)',
-      color: Colors.white,
-    },
-  },
-  controllerToggle: {
-    width: 28,
-    height: 28,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 14,
-    background: 'rgba(0,0,0,0.5)',
-    backdropFilter: 'blur(8px)',
-    border: `1px solid ${Colors.white10}`,
-    color: Colors.white70,
-    cursor: 'pointer',
-    '&:hover': {
-      background: 'rgba(0,0,0,0.7)',
-      color: Colors.white,
-    },
-  },
-  controllerToggleOff: {
-    opacity: 0.4,
-  },
-  actionButton: {
-    width: 32,
-    height: 32,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 8,
-    background: 'rgba(255,255,255,0.12)',
-    border: '1px solid rgba(255,255,255,0.12)',
-    color: Colors.white70,
-    cursor: 'pointer',
-    '&:hover': {
-      background: 'rgba(255,255,255,0.2)',
-      color: Colors.white,
-    },
-  },
-  actionButtonIcon: {
-    fontSize: 18,
-  },
-  actionButtonPortrait: {
-    width: 28,
-    height: 28,
-    borderRadius: 7,
-  },
-  actionButtonIconPortrait: {
-    fontSize: 16,
-  },
-  // Controller overlay
-  controllerOverlay: {
-    position: 'absolute',
-    bottom: 16,
-    right: 16,
-    zIndex: 10,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: '24px',
-    pointerEvents: 'none',
-  },
-  triggerContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    gap: '4px',
-  },
-  triggerShape: {
-    width: 48,
-    height: 80,
-    borderRadius: '8px 8px 24px 24px',
-    border: '2px solid rgba(255,255,255,0.25)',
-    position: 'relative',
-    overflow: 'hidden',
-    background: 'rgba(255,255,255,0.05)',
-    backdropFilter: 'blur(8px)',
-  },
-  triggerFill: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    transition: 'height 50ms linear',
-  },
-  triggerInnerLabel: {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    color: 'rgba(255,255,255,0.4)',
-    zIndex: 1,
-    pointerEvents: 'none',
-  },
-  triggerLabel: {
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    color: 'rgba(255,255,255,0.5)',
-    textTransform: 'uppercase',
-  },
-  controllerJoystick: {
-    width: 100,
-    height: 100,
-    borderRadius: '50%',
-    background: 'rgba(255,255,255,0.05)',
-    backdropFilter: 'blur(8px)',
-    border: '2px solid rgba(255,255,255,0.2)',
-    position: 'relative',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  controllerJoystickTrack: {
-    position: 'absolute',
-    top: '50%',
-    left: 12,
-    right: 12,
-    height: 2,
-    transform: 'translateY(-50%)',
-    background: 'rgba(255,255,255,0.1)',
-    borderRadius: 1,
-  },
-  controllerJoystickThumb: {
-    width: 32,
-    height: 32,
-    borderRadius: '50%',
-    background: 'radial-gradient(circle at 35% 35%, rgba(255,255,255,0.4), rgba(255,255,255,0.1))',
-    boxShadow: 'inset 0 1px 4px rgba(255,255,255,0.3), 0 2px 8px rgba(0,0,0,0.3)',
-    border: '1px solid rgba(255,255,255,0.25)',
-    position: 'absolute',
-    transition: 'left 50ms linear',
-  },
-  controllerJoystickLabel: {
-    fontSize: 10,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    color: 'rgba(255,255,255,0.5)',
-    textTransform: 'uppercase',
-    whiteSpace: 'nowrap',
-  },
-  controllerJoystickArrows: {
-    position: 'absolute',
-    top: '50%',
-    transform: 'translateY(-55%)',
-    fontSize: 14,
-    color: 'rgba(255,255,255,0.2)',
-    userSelect: 'none',
-  },
-  bumperShape: {
-    width: 48,
-    height: 24,
-    borderRadius: '12px 12px 4px 4px',
-    border: '2px solid rgba(255,255,255,0.25)',
-    background: 'rgba(255,255,255,0.05)',
-    backdropFilter: 'blur(8px)',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    transition: 'background 100ms, border-color 100ms',
-  },
-  bumperActive: {
-    background: 'rgba(255,255,255,0.25)',
-    borderColor: 'rgba(255,255,255,0.5)',
-  },
-  bumperLabel: {
-    fontSize: 9,
-    fontWeight: 700,
-    letterSpacing: 0.5,
-    color: 'rgba(255,255,255,0.5)',
-    textTransform: 'uppercase',
-    transition: 'color 100ms',
-  },
-  bumperLabelActive: {
-    color: 'rgba(255,255,255,0.9)',
-  },
-  '@keyframes pulse': {
-    '0%, 100%': { opacity: 1 },
-    '50%': { opacity: 0.5 },
-  },
-});
 
 class BodyTeleop extends Component {
   constructor(props) {
@@ -1194,7 +372,11 @@ class BodyTeleop extends Component {
     dx = Math.max(-1, Math.min(1, dx));
     dy = Math.max(-1, Math.min(1, dy));
     this.setState({ thumbPos: { x: dx, y: dy } });
-    this.setFlippedJoystick(dy, -dx);
+    // Power curve (exponent 2) for finer control near center
+    const expo = 2;
+    const cx = Math.sign(dx) * Math.pow(Math.abs(dx), expo);
+    const cy = Math.sign(dy) * Math.pow(Math.abs(dy), expo);
+    this.setFlippedJoystick(cy, -cx);
   }
 
   resetJoystick() {
@@ -1473,10 +655,6 @@ class BodyTeleop extends Component {
     }, 200);
   }
 
-  isMobile() {
-    return isMobile();
-  }
-
   getStatusDotColor() {
     const { connectionState } = this.state;
     let color;
@@ -1487,26 +665,6 @@ class BodyTeleop extends Component {
       default: color = Colors.grey400;
     }
     return color;
-  }
-
-  renderConnectionIndicator() {
-    const { classes } = this.props;
-    const { connectionState } = this.state;
-
-    return (
-      <div className={classes.hudBottomCenter}>
-        <div className={classes.hudPill}>
-          <div
-            className={classes.statusDot}
-            style={{
-              backgroundColor: this.getStatusDotColor(),
-              animation: connectionState === 'connecting' ? 'pulse 1.5s ease-in-out infinite' : 'none',
-            }}
-          />
-          <span className={classes.hudText}>{connectionState}</span>
-        </div>
-      </div>
-    );
   }
 
   renderPortraitStatusBar() {
@@ -1586,27 +744,24 @@ class BodyTeleop extends Component {
 
     return (
       <div className={classes.hudTopRight}>
-        {connected && (
-          <div
-            className={classes.statsToggleButton}
-            onClick={() => {
-                  this.setState((prev) => {
-                    const next = !prev.showStats;
-                    this.connection.setTimingSei(next);
-                    return { showStats: next };
-                  });
-                }}
-            title="Toggle stats"
-          >
-            STATS
-          </div>
-        )}
+        <div
+          className={classes.hudPillButton}
+          onClick={() => {
+                this.setState((prev) => {
+                  const next = !prev.showStats;
+                  this.connection.setTimingSei(next);
+                  return { showStats: next };
+                });
+              }}
+          title="Toggle stats"
+        >
+          <span className={classes.hudText}>stats</span>
+        </div>
         <div className={classes.hudPill}>
           <div
             className={classes.statusDot}
             style={{
               backgroundColor: this.getStatusDotColor(),
-              animation: connectionState === 'connecting' ? 'pulse 1.5s ease-in-out infinite' : 'none',
             }}
           />
           <span className={classes.hudText}>{connectionState}</span>
@@ -1639,7 +794,8 @@ class BodyTeleop extends Component {
                   {CAMERAS.map((cam) => (
                     <div
                       key={cam.key}
-                      className={`${classes.cameraButton} ${classes.cameraButtonPortrait} ${activeCamera === cam.key ? classes.cameraButtonActive : classes.cameraButtonInactive}`}
+                      className={`${classes.controlButton} ${classes.controlButtonPortrait} ${activeCamera === cam.key ? classes.controlButtonActive : classes.controlButtonInactive}`}
+                      style={{aspectRatio:"1/1"}}
                       onClick={() => this.switchCamera(cam.key)}
                     >
                       {cam.num}
@@ -1655,7 +811,7 @@ class BodyTeleop extends Component {
                 {QUICK_SOUNDS.map((sound) => (
                   <div
                     key={sound.key}
-                    className={`${classes.soundButton} ${classes.soundButtonPortrait}`}
+                    className={`${classes.controlButton} ${classes.controlButtonPortrait}`}
                     onClick={() => this.handlePlaySound(sound.key)}
                   >
                     {sound.label}
@@ -1677,10 +833,9 @@ class BodyTeleop extends Component {
               {CAMERAS.map((cam) => (
                 <div
                   key={cam.key}
-                  className={`${classes.cameraButton} ${activeCamera === cam.key ? classes.cameraButtonActive : classes.cameraButtonInactive}`}
+                  className={`${classes.controlButton} ${activeCamera === cam.key ? classes.controlButtonActive : classes.controlButtonInactive}`}
                   onClick={() => this.switchCamera(cam.key)}
                 >
-                  <span className={classes.cameraButtonKey}>{cam.num}</span>
                   {cam.label}
                 </div>
               ))}
@@ -1693,7 +848,7 @@ class BodyTeleop extends Component {
             {QUICK_SOUNDS.map((sound) => (
               <div
                 key={sound.key}
-                className={classes.soundButton}
+                className={classes.controlButton}
                 onClick={() => this.handlePlaySound(sound.key)}
               >
                 {sound.label}
@@ -1705,14 +860,14 @@ class BodyTeleop extends Component {
         <div className={classes.controlsColumn}>
           <div className={classes.controlsButtons}>
             <div
-              className={`${classes.actionButton} ${streamMuted ? classes.controllerToggleOff : ''}`}
+              className={`${classes.controlButton} ${streamMuted ? classes.controllerToggleOff : ''}`}
               onClick={this.toggleStreamMuted}
               title={streamMuted ? 'Unmute stream audio' : 'Mute stream audio'}
             >
               {streamMuted ? <VolumeOff className={classes.actionButtonIcon} /> : <VolumeUp className={classes.actionButtonIcon} />}
             </div>
             <div
-              className={`${classes.actionButton} ${micMuted ? classes.controllerToggleOff : ''}`}
+              className={`${classes.controlButton} ${micMuted ? classes.controllerToggleOff : ''}`}
               onClick={this.handleMicToggle}
               title={micTitle}
               style={{ display: 'none', pointerEvents: 'none', opacity: 0.5 }}
@@ -1726,7 +881,7 @@ class BodyTeleop extends Component {
         <div className={classes.controlsColumn}>
           <div className={classes.controlsButtons}>
             <div
-              className={classes.actionButton}
+              className={classes.controlButton}
               onClick={() => this.handleScreenshot()}
               title="Save screenshot"
             >
@@ -1835,34 +990,6 @@ class BodyTeleop extends Component {
       </div>
     );
   }
-
-  renderWasdKeys() {
-    const { classes } = this.props;
-    const { keys } = this.state;
-
-    const WasdKey = ({ label, active, keyName }) => (
-      <div
-        className={`${classes.wasdKey} ${active ? classes.wasdKeyActive : classes.wasdKeyInactive}`}
-        onMouseDown={() => this.setKey(keyName, true)}
-        onMouseUp={() => this.setKey(keyName, false)}
-        onMouseLeave={() => active && this.setKey(keyName, false)}
-      >
-        {label}
-      </div>
-    );
-
-    return (
-      <div className={classes.wasdContainer}>
-        <WasdKey label="W" active={keys.w} keyName="w" />
-        <div className={classes.wasdRow}>
-          <WasdKey label="A" active={keys.a} keyName="a" />
-          <WasdKey label="S" active={keys.s} keyName="s" />
-          <WasdKey label="D" active={keys.d} keyName="d" />
-        </div>
-      </div>
-    );
-  }
-
 
   drawLatencyGraph() {
     const canvas = this.latencyCanvasRef.current;
@@ -2025,103 +1152,6 @@ class BodyTeleop extends Component {
     );
   }
 
-  handleOpenTrustPage() {
-    const { directAddress } = this.props;
-    const trustUrl = `${getDeviceBaseUrl(directAddress)}/trust`;
-    window.open(trustUrl, '_blank');
-  }
-
-  getSslInstructions() {
-    if (isSafari()) {
-      return [
-        'Make sure Body ignition is ON',
-        'Click "Open Trust Page"',
-        'Tap "Show Details"',
-        'Tap "visit this website"',
-        'Tap "Visit Website" to confirm',
-      ];
-    }
-    if (isFirefox()) {
-      return [
-        'Make sure Body ignition is ON',
-        'Click "Open Trust Page"',
-        'Click "Advanced\u2026"',
-        'Click "Accept Risk and Continue"',
-      ];
-    }
-    if (isChrome() && isIos()) {
-      return [
-        'Make sure Body ignition is ON',
-        'Click "Open Trust Page"',
-        'Tap "Advanced"',
-        'Tap "Proceed to ... (unsafe)"',
-      ];
-    }
-    // Chrome desktop/Android
-    return [
-      'Make sure Body ignition is ON',
-      'Click "Open Trust Page"',
-      'Click "Advanced"',
-      'Click "Proceed to ... (unsafe)"',
-    ];
-  }
-
-  renderSslTrustDialog() {
-    const { classes, directAddress } = this.props;
-    const { isLandscape } = this.state;
-    const trustUrl = `${getDeviceBaseUrl(directAddress)}/trust`;
-    const steps = this.getSslInstructions();
-
-    return (
-      <div className={classes.connectOverlay} style={{ overflow: 'auto' }}>
-        <div className={classes.connectContent} style={{
-          maxWidth: isLandscape ? 520 : 380,
-          width: '90%',
-          background: Colors.grey900,
-          borderRadius: 16,
-          padding: isLandscape ? 16 : 24,
-          gap: isLandscape ? 10 : 16,
-          margin: 'auto',
-        }}>
-          <Typography style={{ fontSize: isLandscape ? 16 : 18, fontWeight: 600 }}>
-            Trust Device Certificate
-          </Typography>
-          <Typography style={{ fontSize: isLandscape ? 12 : 14, color: Colors.white60, lineHeight: 1.5, textAlign: 'center' }}>
-            Your browser needs to trust this local network device&apos;s self-signed cert
-            before connecting. Follow these steps:
-          </Typography>
-          <div style={{
-            textAlign: 'left',
-            background: Colors.white08,
-            borderRadius: 12,
-            padding: isLandscape ? '8px 12px' : '12px 16px',
-            width: '100%',
-          }}>
-            {steps.map((step, i) => (
-              <div key={i} style={{ display: 'flex', alignItems: 'baseline', gap: 8, marginBottom: i < steps.length - 1 ? (isLandscape ? 2 : 6) : 0 }}>
-                <span style={{ fontSize: isLandscape ? 12 : 14, color: Colors.white40, fontWeight: 600, minWidth: 16 }}>{i + 1}.</span>
-                <span style={{ fontSize: isLandscape ? 12 : 14, color: Colors.white70, lineHeight: 1.4 }}>{step}</span>
-              </div>
-            ))}
-          </div>
-          <Button
-            className={classes.connectButton}
-            onClick={() => this.handleOpenTrustPage()}
-            disableRipple
-          >
-            Open Trust Page
-          </Button>
-          <Typography style={{ fontSize: isLandscape ? 10 : 12, color: Colors.white40, wordBreak: 'break-all', textAlign: 'center' }}>
-            {trustUrl}
-          </Typography>
-          <Typography style={{ fontSize: isLandscape ? 12 : 14, color: Colors.white70 }}>
-            After accepting, this page will connect automatically.
-          </Typography>
-        </div>
-      </div>
-    );
-  }
-
   renderConnectOverlay() {
     const { classes } = this.props;
     const { connectionState, error, statusMessage, connectProgress } = this.state;
@@ -2181,13 +1211,18 @@ class BodyTeleop extends Component {
             </div>
           </div>
 
-          {this.renderHud()}
-          {connected && this.renderStatsOverlay()}
-          {this.state.showSslTrust ? this.renderSslTrustDialog() : !connected && this.renderConnectOverlay()}
-          {connected && this.state.gamepadConnected
-            ? this.renderControllerOverlay()
-            : connected && this.renderJoystick()}
-          {connected && this.renderControls(false)}
+          { connected ? (
+            <>
+              {this.renderHud()}
+              {this.renderStatsOverlay()}
+              {this.state.gamepadConnected
+                ? this.renderControllerOverlay()
+                : this.renderJoystick()}
+              {this.renderControls(false)}
+            </>
+          ) : (
+            this.renderConnectOverlay()
+          )}
         </div>
       </div>
     );
@@ -2222,64 +1257,59 @@ class BodyTeleop extends Component {
             <audio ref={this.audioRef} autoPlay muted={this.state.streamMuted} />
             {connected && this.renderStatsOverlay(true)}
           </div>
-          {connected && this.renderControls(true)}
-          {connected && this.state.gamepadConnected
-            ? this.renderControllerOverlay()
-            : connected && (
+          {connected ?
+            <>
+              {this.renderControls(true)}
               <div style={{ flex: 1, display: 'flex', alignItems: 'center', justifyContent: 'center', padding: 8, minHeight: 0, overflow: 'hidden' }}>
                 {this.renderJoystick(true)}
               </div>
-            )}
-          {!connected && (
+            </>
+          :
             <div className={classes.portraitContent} style={{ overflow: 'auto' }}>
-              {this.state.showSslTrust ? this.renderSslTrustDialog() : (
-                <>
-                  <div className={classes.statusRow}>
-                    <div className={classes.statusLeft}>
-                      <div className={classes.statusDot} style={{ backgroundColor: this.getStatusDotColor(), width: 10, height: 10 }} />
-                      <Typography style={{ fontSize: 14, textTransform: 'capitalize' }}>{connectionState}</Typography>
-                    </div>
-                    {batteryLevel !== null && (
-                      <div className={classes.batteryPill}>
-                        <BatteryFull style={{ fontSize: 18, color: Colors.white70 }} />
-                        <Typography style={{ fontSize: 14 }}>{batteryLevel}%</Typography>
-                      </div>
-                    )}
+              <div className={classes.statusRow}>
+                <div className={classes.statusLeft}>
+                  <div className={classes.statusDot} style={{ backgroundColor: this.getStatusDotColor(), width: 10, height: 10 }} />
+                  <Typography style={{ fontSize: 14, textTransform: 'capitalize' }}>{connectionState}</Typography>
+                </div>
+                {batteryLevel !== null && (
+                  <div className={classes.batteryPill}>
+                    <BatteryFull style={{ fontSize: 18, color: Colors.white70 }} />
+                    <Typography style={{ fontSize: 14 }}>{batteryLevel}%</Typography>
                   </div>
-                  {error && (
-                    <div style={{ borderRadius: 8, background: 'rgba(220,38,38,0.15)', padding: 12, fontSize: 14, color: '#fca5a5' }}>
-                      {error}
-                    </div>
-                  )}
-                  {connecting ? (
-                    <>
-                      <div className={classes.progressBar} style={{ width: '100%' }}>
-                        <div className={classes.progressFill} style={{ width: `${this.state.connectProgress || 0}%` }} />
-                      </div>
-                      <Typography style={{ fontSize: 12, color: Colors.white50, textAlign: 'center' }}>
-                        {this.state.statusMessage || 'Connecting...'}
-                      </Typography>
-                    </>
-                  ) : connectionState === 'failed' ? (
-                    <Button
-                      variant="contained"
-                      className={classes.portraitButton}
-                      style={{ background: Colors.red400, color: Colors.white }}
-                      onClick={this.handleConnect}
-                    >
-                      Retry
-                    </Button>
-                  ) : null}
-                  <div className={classes.infoBox}>
-                    <div className={classes.infoRow}>
-                      <InfoOutline style={{ fontSize: 18 }} />
-                      <span>The comma body must be powered on and ignition must be started to connect.</span>
-                    </div>
-                  </div>
-                </>
+                )}
+              </div>
+              {error && (
+                <div style={{ borderRadius: 8, background: 'rgba(220,38,38,0.15)', padding: 12, fontSize: 14, color: '#fca5a5' }}>
+                  {error}
+                </div>
               )}
+              {connecting ? (
+                <>
+                  <div className={classes.progressBar} style={{ width: '100%' }}>
+                    <div className={classes.progressFill} style={{ width: `${this.state.connectProgress || 0}%` }} />
+                  </div>
+                  <Typography style={{ fontSize: 12, color: Colors.white50, textAlign: 'center' }}>
+                    {this.state.statusMessage || 'Connecting...'}
+                  </Typography>
+                </>
+              ) : connectionState === 'failed' ? (
+                <Button
+                  variant="contained"
+                  className={classes.portraitButton}
+                  style={{ background: Colors.red400, color: Colors.white }}
+                  onClick={this.handleConnect}
+                >
+                  Retry
+                </Button>
+              ) : null}
+              <div className={classes.infoBox}>
+                <div className={classes.infoRow}>
+                  <InfoOutline style={{ fontSize: 18 }} />
+                  <span>The comma body must be powered on and ignition must be started to connect.</span>
+                </div>
+              </div>
             </div>
-          )}
+          }
           {connected && (
             <div style={{ flexShrink: 0, padding: '8px 16px 16px' }}>
               <Button
